// src/app/models/admin/capphephoatdongkhoangsan/cpkhaithackhoangsan/cpkhaithackhuvuc.model.ts
export class OutputCpKhaiThacKhuVucModel {
    idkhaithackhuvuc: string;
    tenkhuvuc: string;
    dientich: number;
    donvidientich: string;
    iddangkykhaithac: string;
    loaikhuvuc: number;
    loaicapphep: string;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    siteid: string;
    listtoado: any;
}

export class InputCpKhaiThacKhuVucModel {
    idkhaithackhuvuc: string;
    tenkhuvuc: string;
    dientich: number;
    donvidientich: string;
    idcapphepkhaithac: string;
    loaikhuvuc: number;
    loaicapphep: string;
    hequychieu: string;
    listtoado: any;
}